discovered new glitch, added audio, added powerup, improved level design, added health items

+++ Assets/scripts/EnemyScripts/States/EnemyHealthBar.cs
@@ -13,12 +13,12 @@ public class EnemyHealthBar : MonoBehaviour
         enemyScript = enemy.GetComponent<EnemyScript>();
     }
 
-    public void SetMaxHealth(int health)
+    public void SetMaxHealth(float health)
     {
         slider.maxValue = health;
 
     }
-    public void UpdateHealthBar(int health)
+    public void UpdateHealthBar(float health)
     {
         slider.value = health;
     }
